moved heartbeat from router to model.
